backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.4
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
passlib==1.7.4
proto-plus==1.27.1
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
